AzureFunc/requirements.txt
azure-functions
Pillow
pdfkit
weasyprint
docx2pdf